# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Authentication and Authorization Micro-service for the FamTrust Web App.",
        "title": "FamTrust API Backend - Auth",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/verify-bvn": {
            "get": {
                "description": "Verify User Signup BVN - Currently this is any 10 digit positive integer not currently in use by another user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications"
                ],
                "summary": "Verify User Signup BVN",
                "operationId": "verify-bvn",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "BVN",
                        "name": "bvn",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/verify-email": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send User-Email Verification Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications"
                ],
                "summary": "Send User-Email Verification Token",
                "operationId": "send-verify-token",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/verify-email/verify": {
            "get": {
                "description": "Verify User Email Address via Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications"
                ],
                "summary": "Verify User Email Address via Token",
                "operationId": "verify-email-token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email verification Token",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/verify-nin": {
            "get": {
                "description": "Verify User Signup NIN - Currently this is any 10 digit positive integer not currently in use by another user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verifications"
                ],
                "summary": "Verify User Signup NIN",
                "operationId": "verify-nin",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "NIN",
                        "name": "nin",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}